+++ htof/special_parse.py
@@ -5,7 +5,7 @@ import pkg_resources
 from astropy.time import Time
 from astropy.table import Table
 from htof.fit import AstrometricFitter
-from htof.parse import HipparcosRereductionJavaTool
+from htof.parse import HipparcosRereductionJavaTool, DataParser
 
 
 class Hipparcos2Recalibrated(HipparcosRereductionJavaTool):
@@ -47,12 +47,12 @@ class Hipparcos2Recalibrated(HipparcosRereductionJavaTool):
             # apply the calibrations
             self.residuals += self.residual_offset
             self.along_scan_errs = np.sqrt(self.along_scan_errs**2 + self.cosmic_dispersion**2)
-            # refit the data, calculate the new residuals, and parameters.
             self.calculate_inverse_covariance_matrices()
-            pf = self.parallax_factors
-            ra_pert = None
-            dec_pert = None
-            parallactic_perturbations = {'ra_plx': ra_pert, 'dec_plx': dec_pert}
+            # munge the parallax factors into the correct form. Note that we are using
+            # the parallax factors from the catalog here to keep everything consistent.
+            ra_motion, dec_motion = to_ra_dec_basis(self.parallax_factors.values, self.scan_angle.values)
+            parallactic_perturbations = {'ra_plx': ra_motion, 'dec_plx': dec_motion}
+            # refit the data, calculate the new residuals, and parameters.
             fitter = AstrometricFitter(inverse_covariance_matrices=self.inverse_covariance_matrix,
                                        epoch_times=Time(Time(self.julian_day_epoch(), format='jd'), format='jyear').value,
                                        central_epoch_dec=1991.25,
@@ -67,17 +67,33 @@ class Hipparcos2Recalibrated(HipparcosRereductionJavaTool):
         return header, raw_data
 
 
-def along_scan_residuals(ra_residual, dec_residual, scan_angle):
+def to_ra_dec_basis(value, scan_angle):
     """
-    Convert residuals in RA and Dec to a residual along the direction of the scan. I.e convert from RA, DEC to
-    Along-scan , Across-scan basis, then keep only the along-scan component.
+    Convert values (e.g. residuals) along the direction of the scan to the same value in RA and Dec. I.e. assume the value
+    has a zero across_scan component and all of the value is in the along-scan direction, then convert that value to a vector in
+    RA and Dec.
     these maths are just from https://en.wikipedia.org/wiki/Rotation_of_axes  .
     """
-    along_scan_residual = dec_residual * np.cos(scan_angle) + ra_residual * np.sin(scan_angle)
-    return along_scan_residual
+    dec_value, ra_value = value * np.cos(scan_angle), value * np.sin(scan_angle)
+    return ra_value, dec_value
+    #return parallax_factor*np.sin(scan_angle), parallax_factor*np.cos(scan_angle)
 
 
-def calculate_new_residuals(data: Hipparcos2Recalibrated, depoch, dplx, dra, ddec, dmura, dmudec):
+def to_along_scan_basis(ra_value, dec_value, scan_angle):
+    """
+    Convert values (e.g. residuals) in RA and Dec to the same value along the direction of the scan. I.e convert from RA, DEC to
+    Along-scan , Across-scan basis, then keep only the along-scan component.
+    these maths are just from https://en.wikipedia.org/wiki/Rotation_of_axes  .
+
+    note that the inputs ra_value and dec_value may have nonzero across scan components. This function will zero
+    out the cross-scan component. to_ra_dec_basis and to_along_scan_basis() are ONLY inverse transforms for data that
+    has components solely in the across-scan direction.
+    """
+    along_scan_value = dec_value * np.cos(scan_angle) + ra_value * np.sin(scan_angle)
+    return along_scan_value
+
+
+def calculate_new_residuals(data: DataParser, depoch, dplx, dra, ddec, dmura, dmudec):
     """
     Recomputes new along-scan residuals, expressed against a perturbed five-parameter catalog solution. If
     the catalog parallax is plx, the catalog ra and dec are ra and dec, and the catalog proper motions are mudec and mura,
@@ -93,5 +109,5 @@ def calculate_new_residuals(data: Hipparcos2Recalibrated, depoch, dplx, dra, dde
     """
     ra_residuals = data.residuals * np.sin(data.scan_angle) - dra - dmura * depoch
     dec_residuals = data.residuals * np.cos(data.scan_angle) - ddec - dmudec * depoch
-    new_residuals = along_scan_residuals(ra_residuals, dec_residuals, data.scan_angle) - dplx * data.parallax_factors
+    new_residuals = to_along_scan_basis(ra_residuals, dec_residuals, data.scan_angle) - dplx * data.parallax_factors
     return new_residuals